+++ .vscode/settings.json
@@ -54,6 +54,7 @@
     "mitigations",
     "mixtral",
     "MMLU",
+    "moderations",
     "multibar",
     "ngrams",
     "nvmrc",
