+++ src/providers/openai/chat.ts
@@ -8,58 +8,9 @@ import type { EnvOverrides } from '../../types/env';
 import { renderVarsInObject } from '../../util';
 import { maybeLoadFromExternalFile } from '../../util';
 import { REQUEST_TIMEOUT_MS, parseChatPrompt } from '../shared';
-import type { OpenAiSharedOptions } from './types';
+import type { OpenAiCompletionOptions } from './types';
 import { calculateOpenAICost } from './util';
-import {
-  formatOpenAiError,
-  getTokenUsage,
-  OPENAI_CHAT_MODELS,
-  type OpenAiFunction,
-  type OpenAiTool,
-} from './util';
-
-export type OpenAiCompletionOptions = OpenAiSharedOptions & {
-  temperature?: number;
-  max_completion_tokens?: number;
-  max_tokens?: number;
-  top_p?: number;
-  frequency_penalty?: number;
-  presence_penalty?: number;
-  best_of?: number;
-  functions?: OpenAiFunction[];
-  function_call?: 'none' | 'auto' | { name: string };
-  tools?: OpenAiTool[];
-  tool_choice?: 'none' | 'auto' | 'required' | { type: 'function'; function?: { name: string } };
-  response_format?:
-    | {
-        type: 'json_object';
-      }
-    | {
-        type: 'json_schema';
-        json_schema: {
-          name: string;
-          strict: boolean;
-          schema: {
-            type: 'object';
-            properties: Record<string, any>;
-            required?: string[];
-            additionalProperties: false;
-          };
-        };
-      };
-  stop?: string[];
-  seed?: number;
-  passthrough?: object;
-
-  /**
-   * If set, automatically call these functions when the assistant activates
-   * these function tools.
-   */
-  functionToolCallbacks?: Record<
-    OpenAI.FunctionDefinition['name'],
-    (arg: string) => Promise<string>
-  >;
-};
+import { formatOpenAiError, getTokenUsage, OPENAI_CHAT_MODELS } from './util';
 
 export class OpenAiChatCompletionProvider extends OpenAiGenericProvider {
   static OPENAI_CHAT_MODELS = OPENAI_CHAT_MODELS;
