+++ src/providers/openai/index.ts
@@ -1,4 +1,3 @@
-import type OpenAI from 'openai';
 import { getEnvString } from '../../envars';
 import type {
   ApiProvider,
@@ -8,50 +7,6 @@ import type {
 } from '../../types';
 import type { EnvOverrides } from '../../types/env';
 import type { OpenAiSharedOptions } from './types';
-import { type OpenAiFunction, type OpenAiTool } from './util';
-
-export type OpenAiCompletionOptions = OpenAiSharedOptions & {
-  temperature?: number;
-  max_completion_tokens?: number;
-  max_tokens?: number;
-  top_p?: number;
-  frequency_penalty?: number;
-  presence_penalty?: number;
-  best_of?: number;
-  functions?: OpenAiFunction[];
-  function_call?: 'none' | 'auto' | { name: string };
-  tools?: OpenAiTool[];
-  tool_choice?: 'none' | 'auto' | 'required' | { type: 'function'; function?: { name: string } };
-  response_format?:
-    | {
-        type: 'json_object';
-      }
-    | {
-        type: 'json_schema';
-        json_schema: {
-          name: string;
-          strict: boolean;
-          schema: {
-            type: 'object';
-            properties: Record<string, any>;
-            required?: string[];
-            additionalProperties: false;
-          };
-        };
-      };
-  stop?: string[];
-  seed?: number;
-  passthrough?: object;
-
-  /**
-   * If set, automatically call these functions when the assistant activates
-   * these function tools.
-   */
-  functionToolCallbacks?: Record<
-    OpenAI.FunctionDefinition['name'],
-    (arg: string) => Promise<string>
-  >;
-};
 
 export class OpenAiGenericProvider implements ApiProvider {
   modelName: string;
