+++ src/providers/openai/types.ts
@@ -1,3 +1,6 @@
+import type OpenAI from 'openai';
+import { type OpenAiFunction, type OpenAiTool } from './util';
+
 export interface OpenAiSharedOptions {
   apiKey?: string;
   apiKeyEnvar?: string;
@@ -7,3 +10,46 @@ export interface OpenAiSharedOptions {
   cost?: number;
   headers?: { [key: string]: string };
 }
+
+export type OpenAiCompletionOptions = OpenAiSharedOptions & {
+  temperature?: number;
+  max_completion_tokens?: number;
+  max_tokens?: number;
+  top_p?: number;
+  frequency_penalty?: number;
+  presence_penalty?: number;
+  best_of?: number;
+  functions?: OpenAiFunction[];
+  function_call?: 'none' | 'auto' | { name: string };
+  tools?: OpenAiTool[];
+  tool_choice?: 'none' | 'auto' | 'required' | { type: 'function'; function?: { name: string } };
+  response_format?:
+    | {
+        type: 'json_object';
+      }
+    | {
+        type: 'json_schema';
+        json_schema: {
+          name: string;
+          strict: boolean;
+          schema: {
+            type: 'object';
+            properties: Record<string, any>;
+            required?: string[];
+            additionalProperties: false;
+          };
+        };
+      };
+  stop?: string[];
+  seed?: number;
+  passthrough?: object;
+
+  /**
+   * If set, automatically call these functions when the assistant activates
+   * these function tools.
+   */
+  functionToolCallbacks?: Record<
+    OpenAI.FunctionDefinition['name'],
+    (arg: string) => Promise<string>
+  >;
+};
